Fragile checkpoint: plot working.
Found mistake! Inline definitions of React functional components compare
to false because their references are different.

React uses the reference of the class/function to diff on the type of
component. Even if the ASCII letters look hardcoded as the same, if the
definition is changing on every render, then indeed so is the type.

My inline definition of the PusherData HOC was the issue. It was being
defined in the App.render function.

This had nothing to do with the React Context API being strange. It
works just fine.

tl;dr don't expect inline Component definitions to persist and remember
that React unmounts a component if its type changes.